Change the contract of get_typesupport_handle_function.
Before it would set an RCUTILS error if the typesupport
didn't match.  But because this is essentially a probe, we
change it to not set that on error and instead just return
a nullptr.  Callers can disambiguate between the probe not
finding a typesupport and a real error by checking the return
value for nullptr and then checking if `rcutils_error_is_set`.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>