Make sure to check typesupport handles against nullptr properly.
clang has a false-positive warning where it can't figure out
that ASSERT_TRUE(foo != nullptr) means that the rest of the
test is safe to dereference foo.  Switching to ASSERT_NE(foo, nullptr)
fixes the issue, so do that here.  This should get rid of a
number of warnings when building under clang static analysis.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>